update types and imports
